Move PluginPriority to extension_points
- it is used mainly as an optional parameter by register_*() extension methods
- make it an IntEnum
